tp1/ship_classification/requirements.txt
Pillow==10.2.0
joblib==1.3.2
fastapi[standard]
scikit-learn==1.4.1.post1